src/faces/face-describe-py/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
absl-py==0.8.0            # via tensorboard, tensorflow
astor==0.8.0              # via tensorflow
backports.weakref==1.0.post1  # via tensorflow
cachetools==3.1.1         # via google-auth
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
enum34==1.1.6             # via absl-py, grpcio, tensorflow
funcsigs==1.0.2           # via mock
futures==3.3.0            # via google-api-core, grpcio, tensorboard
gast==0.2.2               # via tensorflow
google-api-core==1.6.0
google-auth==1.6.3        # via google-api-core
google-pasta==0.1.7       # via tensorflow
googleapis-common-protos==1.6.0  # via google-api-core
grpcio==1.16.1
h5py==2.9.0               # via keras-applications
idna==2.8                 # via requests
keras-applications==1.0.8  # via tensorflow
keras-preprocessing==1.1.0  # via tensorflow
markdown==3.1.1           # via tensorboard
mock==3.0.5               # via tensorflow
numpy==1.16.5
pillow==6.1.0
protobuf==3.9.1           # via google-api-core, googleapis-common-protos, tensorboard, tensorflow
pyasn1-modules==0.2.6     # via google-auth
pyasn1==0.4.7             # via pyasn1-modules, rsa
python-dotenv==0.10.3
python-json-logger==0.1.9
pytz==2019.2              # via google-api-core
requests==2.22.0          # via google-api-core
rsa==4.0                  # via google-auth
scipy==1.2.2
six==1.12.0               # via absl-py, google-api-core, google-auth, grpcio, h5py, keras-preprocessing, mock, protobuf, tensorboard, tensorflow
tensorboard==1.14.0       # via tensorflow
tensorflow-estimator==1.14.0  # via tensorflow
tensorflow==1.14.0
termcolor==1.1.0          # via tensorflow
typing==3.7.4.1           # via python-dotenv
urllib3==1.25.3           # via requests
werkzeug==0.15.6          # via tensorboard
wheel==0.33.6             # via tensorboard, tensorflow
wrapt==1.11.2             # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via google-api-core, markdown, protobuf, python-json-logger, tensorboard
